Use UInt64 for the ZSTD frame content size
clang-tidy google-runtime-int rejects the bare `unsigned long long` local
holding ZSTD_getFrameContentSize's result, which fails Build (arm_tidy) since
-Werror is on. uint64_t (UInt64) has the same width, so both sentinel
comparisons against ZSTD_CONTENTSIZE_ERROR and ZSTD_CONTENTSIZE_UNKNOWN keep
their meaning; verified by static_assert on x86-64 and aarch64, and by running
clang-tidy-21 with the repo config over the translation unit.

+++ src/Processors/Formats/Impl/ArrowIPC/BufferCompression.cpp
@@ -36,7 +36,7 @@ std::optional<UInt64> declaredFrameContentSize(CompressionCodec codec, const cha
             throw Exception(ErrorCodes::INCORRECT_DATA, "Compressed Arrow IPC buffer is not a valid ZSTD frame");
         if (frame_size != size)
             return std::nullopt;
-        const unsigned long long n = ZSTD_getFrameContentSize(src, size);
+        const UInt64 n = ZSTD_getFrameContentSize(src, size);
         if (n == ZSTD_CONTENTSIZE_ERROR)
             throw Exception(ErrorCodes::INCORRECT_DATA, "Compressed Arrow IPC buffer is not a valid ZSTD frame");
         if (n == ZSTD_CONTENTSIZE_UNKNOWN)
